+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Master](https://github.com//AppwiseCore)
 
+### New Features
+
+* Core: you can now unregister observers from typed notifications.
+
 ## [0.8.11](https://github.com/appwise-labs/AppwiseCore/releases/tag/0.8.11)
 
 ### New Features